+++ src/main/java/sic/repository/FacturaVentaRepositoryCustom.java
@@ -1,5 +1,6 @@
 package sic.repository;
 
+import java.math.BigDecimal;
 import org.springframework.data.domain.Page;
 import sic.modelo.BusquedaFacturaVentaCriteria;
 import sic.modelo.FacturaVenta;
@@ -7,11 +8,11 @@ import sic.modelo.TipoDeComprobante;
 
 public interface FacturaVentaRepositoryCustom {
 
-    double calcularTotalFacturadoVenta(BusquedaFacturaVentaCriteria criteria);
+    BigDecimal calcularTotalFacturadoVenta(BusquedaFacturaVentaCriteria criteria);
 
-    double calcularIVA_Venta(BusquedaFacturaVentaCriteria criteria, TipoDeComprobante[] tipoComprobante);
+    BigDecimal calcularIVA_Venta(BusquedaFacturaVentaCriteria criteria, TipoDeComprobante[] tipoComprobante);
 
-    double calcularGananciaTotal(BusquedaFacturaVentaCriteria criteria);
+    BigDecimal calcularGananciaTotal(BusquedaFacturaVentaCriteria criteria);
 
     Page<FacturaVenta> buscarFacturasVenta(BusquedaFacturaVentaCriteria criteria);
